chore(clippy): 清零 296 个历史 warning + CI 收紧严门禁
clippy --fix 修机械类（unnecessary_cast/needless_borrow/collapsible_if 等）；
各 crate root 加 #![allow] 显式放行可辩护的测试/FFI 模式 lint 并注明理由：
- core: field_reassign_with_default(测试 setup) / type_complexity / too_many_arguments(管线 fn)
  / neg_cmp_op_on_partial_ord(浮点) / wrong_self_convention(trait 签名) / doc_lazy_continuation
- ffi:  not_unsafe_ptr_arg_deref / too_many_arguments / type_complexity / unnecessary_cast(abi raw ptr)
  / needless_range_loop(blob/tests 顶点填装) / field_reassign_with_default
- pkg:  type_complexity

真错的修了：ffi manual_memcpy→copy_from_slice、3 处 get().is_none()→!contains_key()。

rust-ci.yml clippy job 从 advisory 收紧为 -D warnings（基线零 warning）。
cargo clippy --all-targets -- -D warnings 过；cargo test --workspace 全绿。

+++ loomgui_core/examples/dump_interact.rs
@@ -126,7 +126,7 @@ fn main() {
 
     let mut s = Stage::new(font, (200.0, 100.0)).unwrap();
     s.load_package("showcase", &pkg).unwrap();
-    let btn = s.find_node_by_id("b1").expect("b1").0 as u32;
+    let btn = s.find_node_by_id("b1").expect("b1").0;
     let r = s
         .scene
         .as_ref()
